chore: move `lib` to `utils`
