+++ packages/next-auth/src/core/lib/oauth/authorization-url.ts
@@ -4,7 +4,7 @@ import { createState } from "./state-handler"
 import { createPKCE } from "./pkce-handler"
 
 import type { AuthorizationParameters } from "openid-client"
-import type { InternalOptions } from "../../../lib/types"
+import type { InternalOptions } from "../../types"
 import type { RequestInternal } from "../.."
 import type { Cookie } from "../cookie"
