+++ packages/next-auth/src/core/lib/oauth/client-legacy.ts
@@ -2,7 +2,7 @@
 // We have the intentions to provide only minor fixes for this in the future.
 
 import { OAuth } from "oauth"
-import { InternalOptions } from "src/lib/types"
+import type { InternalOptions } from "../../types"
 
 /**
  * Client supporting OAuth 1.x
