+++ packages/next-auth/src/core/lib/oauth/client.ts
@@ -1,5 +1,6 @@
-import { Issuer, Client, custom } from "openid-client"
-import { InternalOptions } from "src/lib/types"
+import { Issuer, custom } from "openid-client"
+import type { Client } from "openid-client"
+import type { InternalOptions } from "../../types"
 
 /**
  * NOTE: We can add auto discovery of the provider's endpoint
@@ -12,9 +13,9 @@ export async function openidClient(
   options: InternalOptions<"oauth">
 ): Promise<Client> {
   const provider = options.provider
-  
+
   if (provider.httpOptions) custom.setHttpOptionsDefaults(provider.httpOptions)
-  
+
   let issuer: Issuer
   if (provider.wellKnown) {
     issuer = await Issuer.discover(provider.wellKnown)
