+++ packages/next-auth/src/core/pages/index.ts
@@ -5,7 +5,7 @@ import VerifyRequestPage from "./verify-request"
 import ErrorPage from "./error"
 import css from "../../css"
 
-import type { InternalOptions } from "../../lib/types"
+import type { InternalOptions } from "../types"
 import type { RequestInternal, OutgoingResponse } from ".."
 import type { Cookie } from "../lib/cookie"
 import type { ErrorType } from "./error"
