+++ packages/next-auth/src/core/pages/signin.tsx
@@ -1,5 +1,4 @@
-import { Theme } from "../.."
-import { InternalProvider } from "../../lib/types"
+import type { InternalProvider, Theme } from "../types"
 
 /**
  * The following errors are passed as error query parameters to the default or overridden sign-in page.
