+++ packages/next-auth/src/core/routes/session.ts
@@ -1,7 +1,7 @@
 import { fromDate } from "../lib/utils"
 
 import type { Adapter } from "../../adapters"
-import type { InternalOptions } from "../../lib/types"
+import type { InternalOptions } from "../types"
 import type { OutgoingResponse } from ".."
 import type { Session } from "../.."
 import type { SessionStore } from "../lib/cookie"
