+++ packages/next-auth/src/next/utils.ts
@@ -13,11 +13,3 @@ export function setCookie(res, cookie: Cookie) {
   setCookieHeader.push(cookieHeader)
   res.setHeader("Set-Cookie", setCookieHeader)
 }
-
-/** Extract the host from the environment */
-export function detectHost(forwardedHost: any) {
-  // If we detect a Vercel environment, we can trust the host
-  if (process.env.VERCEL) return forwardedHost
-  // If `NEXTAUTH_URL` is `undefined` we fall back to "http://localhost:3000"
-  return process.env.NEXTAUTH_URL
-}
