+++ packages/next-auth/src/utils/merge.ts
